mukayh_stock_system/requirements.txt
# Updated requirements.txt with Python 3.13 compatible versions

# Core Django
Django==5.0.2  # Updated for Python 3.13
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.0

# Data Science & AI/ML Libraries
numpy>=1.26.0  # Updated for Python 3.13
pandas>=2.1.4  # Updated for Python 3.13
scikit-learn>=1.3.2
statsmodels>=0.14.1
prophet>=1.1.5
tensorflow>=2.15.0  # Note: Check TF compatibility with Python 3.13

# Data Visualization
matplotlib>=3.8.2
seaborn>=0.13.0
plotly>=5.18.0

# Task Queue
celery>=5.3.4
redis>=5.0.1
django-celery-beat>=2.5.0

# Excel/CSV Processing
openpyxl>=3.1.2
xlsxwriter>=3.1.9

# Environment Variables
python-decouple>=3.8
python-dotenv>=1.0.0

# Date/Time utilities
python-dateutil>=2.8.2
pytz>=2023.3

# API Documentation
drf-spectacular>=0.27.0

# Testing
pytest>=7.4.3
pytest-django>=4.7.0
factory-boy>=3.3.0
faker>=22.0.0

# Code Quality
flake8>=6.1.0
black>=23.12.0
isort>=5.13.2

# Production Server
gunicorn>=21.2.0
whitenoise>=6.6.0

# Monitoring & Logging
sentry-sdk>=1.39.1
django-extensions>=3.2.3

# Development tools
ipython>=8.18.1
django-debug-toolbar>=4.2.0